refactor: new images placeholder
