+++ presentation/src/main/java/br/dev/singular/overview/presentation/ui/screens/person/PersonDetailsScreen.kt
@@ -138,11 +138,11 @@ private fun PersonImage(
         url = person.profileURL,
         contentScale = ContentScale.Crop,
         previewPainter = person.previewContent,
+        errorDefaultImage = R.drawable.error_profile_placeholder,
         modifier = modifier
             .size(268.dp)
             .clip(CircleShape)
-            .border(),
-        errorDefaultImage = R.drawable.avatar
+            .border()
     )
 }
