--- presentation/src/main/res/drawable/placeholder.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="200dp"
-    android:height="200dp"
-    android:viewportWidth="100"
-    android:viewportHeight="100">
-
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M0,0h100v100h-100z" />
-
-    <path
-        android:strokeColor="#33FFFFFF"
-        android:strokeWidth="0.5"
-        android:pathData="M-100,100 L100,-100 M-90,100 L110,-100 M-80,100 L120,-100
-                          M-70,100 L130,-100 M-60,100 L140,-100 M-50,100 L150,-100
-                          M-40,100 L160,-100 M-30,100 L170,-100 M-20,100 L180,-100
-                          M-10,100 L190,-100 M0,100 L200,-100 M10,100 L210,-100
-                          M20,100 L220,-100 M30,100 L230,-100 M40,100 L240,-100" />
-</vector>
